+++ src/objc/class.rs
@@ -145,7 +145,7 @@ impl Class {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418583-respondstoselector).
     #[inline]
     pub fn responds_to_selector(&self, selector: SEL) -> bool {
-        unsafe { _msg_send_cached![self, respondsToSelector: selector => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, respondsToSelector: selector => BOOL] }.into()
     }
 
     /// Returns `true` if instances of this class implement or inherit a method
@@ -154,7 +154,7 @@ impl Class {
     /// See [documentation](https://developer.apple.com/documentation/objectivec/nsobject/1418555-instancesrespondtoselector).
     #[inline]
     pub fn instances_respond_to_selector(&self, selector: SEL) -> bool {
-        unsafe { _msg_send_cached![self, instancesRespondToSelector: selector => BOOL] != 0 }
+        unsafe { _msg_send_cached![self, instancesRespondToSelector: selector => BOOL] }.into()
     }
 
     /// Returns the name of this class.